Define general purpose data object
The schema already defines json_t type, which can hold any JSON value.
The new data object is an unordered collection of attributes (name/value
pairs, like a Map or Dictionary), which are not defined by the schema.

Define a few new attributes that are required for the Authetentication
and Authorization classes.