Rename Audit to Access Control
Per discussion on 2023-03-14, renames "Audit Activity" category to "Access Control Activity" to remove confusion stemming from generalized definition of "Audit" and bring it closer in line with the present definition of the category: "...events relate to the supervision of the system's access control model. Examples of such events are the success or failure of authentication, granting of authority, password change, entity change, privileged use, system state change, and resource access."

Removes extraneous 'audit.json' from events as this was not being used and is out of line with the proposed adjustment to the name.

Changes path for category and classes within it and changes class association accordingly.

Does not change association of "API Activity" class as was discussed, this should be done in a separate PR once the Application category class is resurrected.