Added 'Windows Service Activity' event. (#1103)
#### Related Issue: 

[1020](https://github.com/ocsf/ocsf-schema/issues/1020)

#### Description of changes:

This schema change introduces a new `Windows Service Activity` event to
the Windows extension in the `System Activity` category. The new event
extends the `System Activity` event only minimally:
* Enumerants are added to the `activity_id` attribute, corresponding to
the various actions that a process may perform on a service.
* The new `win_service` attribute is a `Windows Service` object.

The `Windows Service` object extends the pre-existing `Service` object
in some important ways:
* The pre-existing `name` attribute is made `required` so as to reflect
the mandatory nature of this attribute when describing behaviour
pertaining to a Windows service.
* The new `service_type_id` attribute and its `service_type` sibling
correspond to the `dwServiceType` field in the Win32
`QUERY_SERVICE_CONFIG` structure.
* The new `service_start_type_id` attribute and its `service_start_type`
sibling correspond to the `dwStartType` field in the Win32
`QUERY_SERVICE_CONFIG` structure.
* The new `cmd_line` attribute corresponds to the `lpBinaryPathName`
field in the Win32 `QUERY_SERVICE_CONFIG` structure.
* The new `load_order_group` attribute corresponds to the
`lpLoadOrderGroup` field in the Win32 `QUERY_SERVICE_CONFIG` structure.
* The new `service_dependencies` attribute corresponds to the
`lpDependencies` field in the Win32 `QUERY_SERVICE_CONFIG` structure.
* The new `service_start_name` attribute corresponds to the
`lpServiceStartName` field in the Win32 `QUERY_SERVICE_CONFIG`
structure.
* The new `service_error_control` attribute corresponds to the
`dwErrorControl` field in the Win32 `QUERY_SERVICE_CONFIG` structure.
* The new `service_category_id` attribute and its `service_category`
sibling do not directly correspond to any field in the Win32
`QUERY_SERVICE_CONFIG` structure but instead represent a more
coarse-grained categorisation of the information in the
`service_type_id` and `service_type` attributes. This is useful in
contexts where the precise type of service is unknown and a security
product knows only if the activity relates to a kernel mode driver or a
user mode service.

Because behaviour relating to Windows services is a common trigger for
detections, this PR also adds the `win_service` attribute to the
`Windows Evidence Artifacts` object.

---------

Signed-off-by: Dave McCormack <dmccorma@cisco.com>