Added new metadata fields for cross-referencing via hyperlink, and source (#1189)
#### Related Issue: N/A

#### Description of changes:
Currently any cross-referencing to MITRE or other associated schemas is
done via embedding a link in the description of the object or attribute.
Similarly, OCSF names for attributes that come from well-known sources,
such as LDAP or RFCs aren't formally captured. Two new metadata fields:
`source` and `reference` aim to capture that in the metaschema, with the
intention that the OCSF server will automatically format them and
display them within the appropriate descriptions. This will allow for
programmatic processing by other tools as well.

---------

Signed-off-by: Paul Agbabian <pagbabian@splunk.com>
Co-authored-by: Rick Mouritzen <rmouritzen@splunk.com>
Co-authored-by: Rajas <89877409+floydtree@users.noreply.github.com>