Adding a Tags - key:value pair object to the schema (#1207)
#### Related Issue: From Oct 16th, 2024 Network Activity Call. 
Currently, we have `labels`, a string array attribute that can be used
to populate labels. A String array is not sufficient when labels come in
the form `key:value` pairs. Therefore, adding a tag object which can
allow users to populate key:value styled labels/tags in the events.

#### Description of changes:
1. Adding a `tag` object, adding `tags` to dictionary. 
2. Deprecating an inconsistently used `tag` attribute, captioned `Image
Tag` (was used in
[Image](https://schema.ocsf.io/1.4.0-dev/objects/image?extensions=) and
[Container](https://schema.ocsf.io/1.4.0-dev/objects/container?extensions=))
adding labels and tags to the object as alternatives and for
consistency.
3. Adding `tags` to all the objects where, `labels` is currently used.
And sorting the file contents.
4. Sorting dictionary.
5. Minor description adjustments.

---------

Signed-off-by: Rajas Panat <rajaspa@amazon.com>