Removed redundant attribute from Windows extension to `startup_item` object. (#1203)
#### Issue: 

I raised this issue in Slack and it was discussed at the weekly OCSF
call on 8th October 2024. It was agreed that I would submit this PR.
Apologies for the delay!

A recent PR extended `startup_item` for Windows. The intenion of the
author of that PR was that the extended object would have the same name
as the base object. This is normally achieved by omiting the `name`
attribute from the extended object's definition. However, the `name`
attribute was explicitly specified in that PR. Although this is
permitted by the metaschema, it had the effect of breaking some code
generation tooling. It also meant that the extension was inconsistent
with other non-renaming object extensions.

#### Description of changes:

The `name` attribute has been removed from the Windows extension to the
`startup_item` object.

Signed-off-by: Dave McCormack <dmccorma@cisco.com>
Co-authored-by: Rajas <89877409+floydtree@users.noreply.github.com>