Add `long_string` object for strings that are potentially very long and subject to truncation. (#1228)
#### Related Issue: 

The change being made in this PR was proposed and agreed during the
weekly general call on 29th October. I undertook to implement this
change because it represents a more general way to achieve something
that I had done in a less general way in recents PRs.

#### Description of changes:

This PR introduces a `long_string` object which can be used as a type
for attributes that hold string values that are potentially very long.
Examples of such strings include the content of an executed script, or
the names and values of environment variables. Such strings are likely
to be truncated by security products for reasons of transmission and
storage efficiency.

---------

Signed-off-by: Dave McCormack <dmccorma@cisco.com>
Co-authored-by: Rajas <89877409+floydtree@users.noreply.github.com>