Adding vendor provided severity info to finding_info (#1257)
#### Scenario that this change aims to support

Consider an finding vendor/producer emits a Finding event. The event is
consumed by a different system which allow mutability of events,
effectively, allowing end-users to update values in the finding. Case in
point, consider `severity`. In such a case, how do you ensure the
original vendor/producer's severity value is not lost? -> By utilizing
this object to store the original values. The end-user can the simply
update the top level values in the finding, as the system would maintain
the original values in this object.

#### Description of changes:

1. Adding a new object `vendor_attributes` as an optional attribute to
all Findings category classes (except the deprecated Security Finding
class).
2. Relevant dictionary entries.
3. Fixing missing PR number from a previous changelog entry. #1249 

Object Details - 
![Screenshot 2024-11-20 at 10 58
52](https://github.com/user-attachments/assets/b28988b3-7d19-4a22-97de-38f8b14f20ea)

---------

Signed-off-by: Rajas Panat <rajaspa@amazon.com>