Update Actor description to distinguish from a campaign threat actor (#1290)
#### Related Issue: N/A (Slack discussion)

#### Description of changes:
There was a discussion in Slack that called out some new OCSF users
being confused by `actor` thinking it is a threat actor, i.e. from a
campaign. Although `actor` might be a vehicle for a threat actor, the
descriptions in the dictionary and the object add a sentence making the
distinction.

---------

Signed-off-by: Paul Agbabian <pagbabian@splunk.com>
Co-authored-by: Jonathan Rau <139361268+jonrau-at-queryai@users.noreply.github.com>