Add ancestry array to process object (#1317)
#### Related Issue: 

https://github.com/ocsf/ocsf-schema/issues/1275

#### Description of changes:

Adds an array for the expression of process ancestry through process
entity objects.
The process entity is a new object that the process object is modified
to extend.
This enables the inclusion of extended parentage information without
deep nesting which is discouraged in the `parent_process` attribute
description.
Full details for processes in the ancestry would still need to be looked
up based on the fields in the process entity object.

This PR also deprecates the `lineage` attribute within the process
object in favour of the `ancestry` attribute which has much more
explicit guidance on how to construct the array.

Signed-off-by: Mitchell Wasson <miwasson@cisco.com>