Add 'internal_name' attribute to 'file' object. (#1322)
#### Description of changes:

This PR adds an `internal_name` to the dictionary and to the `file`
object. This attribute is intended to capture the name of the file as
identified within the file itself. This contrasts with the name by which
the file is known on disk.

Where available, the internal name is widely used by security
practitioners and detection content because the on-disk file name is not
reliable.

On the Windows OS, most PE files contain a <code>VERSIONINFO</code>
resource from which the internal name can be obtained.

On macOS, binaries can optionally embed a copy of the application's
Info.plist file which in turn contains the name of the executable
(#NotAMacGuy).

---------

Co-authored-by: Rajas <89877409+floydtree@users.noreply.github.com>