Add raw_data_size to base_event. (#1347)
#### Related Issue: N/A 

This is a follow on [from
slack](https://opencybersecu-lz97379.slack.com/archives/C03C2QPSBPB/p1739819906117879).

#### Description of changes:

These changes add a `raw_data_size` field to `base_event`. This field
can assist in answering questions related to the size of the incoming
data. A key insight would be tracking situations where a system
drastically increases or decreases the average size of a raw message.

While it's possible to derive this calculation today if the underlying
storage system supports counting the number of bytes in a field, it's
not efficient to perform that calculation at scale (worst case: $`O(n
\times m)`$). Adding this field allows for a much more efficient
calculation, and also provides the ability to perform that calculation
in situations where the storage system does not store the `raw_data` or
stores it separate from the OCSF formatted event.

Signed-off-by: Hunter Madison <hunter.madison@ibm.com>