Added count, start_time, end_time to timespan as optional attributes. (#1365)
#### Related Issue: 

#### Description of changes:
Added `start_time` `end_time` and `count` as optional attributes to mark
the duration of the timespan, and count any items measured in the
timespan (if any).

This update followed discussion around baseline and observation windows
for the `Anomaly Analysis` set of objects where a `baseline` and
`observation` object each have windows or timespans. New attributes of
type `timespan` can now be created to specify those windows.

---------

Signed-off-by: Paul Agbabian <pagbabian@splunk.com>