Added Managed Entity Zone type (#1364)
#### Related Issue: Supersedes PR #1181

#### Description of changes:
Added 'Network Zone' to the types of managed entities.

This PR supersedes PR #1181 which can be closed. After group discussion,
there was a conflict with an existing `zone` attribute whose caption and
usage was 'Network Zone', the same type of zone as in PR #1181.

Managed Entity extends `_entity` and has a `name` and `uid` so there was
no need to add another name and uid (e.g. `zone` (existing) and
`zone_uid` (new)). The descriptions have been updated to reflect the
proper usage of the object's `name` and `uid` for use with the 'Network
Zone' type, as well as when other specific types with their own objects
are chosen, for example `user` or `device`, which are entities that have
their own `name` and `uid` which should match the `name` and `uid` of
the Managed Entity object instance.

---------

Signed-off-by: Paul Agbabian <pagbabian@splunk.com>