Added `Detect` value for `activity_id` in Remediation events (#1362)
In Cisco Endpoint Security (and other EDRs), a common response to a
detection is uploading a file to a dynamic analysis sandbox. This action
is covered in the Detect MITRE D3FEND Tactic with [Dynamic
Analysis](https://d3fend.mitre.org/technique/d3f:DynamicAnalysis/).

This PR adds Detect as an `activity_id` in Remediation activity to
accurately model the above situation (and others).

Signed-off-by: Mitchell Wasson <miwasson@cisco.com>
Signed-off-by: Irakle Dzneladze <106536760+irakledibm@users.noreply.github.com>
Signed-off-by: Jonathan Rau <139361268+jonrau-at-queryai@users.noreply.github.com>
Co-authored-by: Rajas <89877409+floydtree@users.noreply.github.com>
Co-authored-by: Irakle Dzneladze <106536760+irakledibm@users.noreply.github.com>
Co-authored-by: Jonathan Rau <139361268+jonrau-at-queryai@users.noreply.github.com>