Add several new observable types (simple cases) (#1380)
This breaks-out simple cases from the
https://github.com/ocsf/ocsf-schema/pull/1326

#### Related Issue: N/A

#### Description of changes:

We are populating the observables with values which are notable for
correlation or display to the users. There are several values, which we
think make good observables, but are not marked as such. This PR tries
to add them.

These are the details:

  1. Added `process_entity.uid` as an Observable type - `type_id: 39`.
1. Added `email.subject` and `email.uid` as an Observable types -
`type_id: 40` and `type_id: 41`.
  1. Added 'message_uid' as Observable type - `type_id: 42`.
  1. Added `reg_value.name` as an Observable type - `type_id: 43`.

I have updated CHANGELOG.md with the same granularity of the information
it already uses, but it's not a "single-line" description.

---------

Signed-off-by: Michal Dobisek <mdobisek@cisco.com>