Add new observables (type change) (#1381)
This breaks-out cases with type-change from the
https://github.com/ocsf/ocsf-schema/pull/1326 - please see the
discussion there for the concerns.

#### Related Issue: N/A

#### Description of changes:

We are populating the observables with values which are notable for
correlation or display to the users. There are several values, which we
think make good observables, but are not marked as such. However, these
fields already use overly generic `path` dictionary attribute. To
selectively mark some of them observables, this PR first introduces more
fine grained types for some `path` attributes.

These are the details:

1. Added `file_path_t` as an Observable type - `type_id: 44` and marked
fields as this type
    - `lineage` dictionary attribute
    - `affected_package.path` object attribute
    - `file.path` object attribute
    - `image.path` object attribute
    - `kernel.path` object attribute
    - `malware.path` object attribute
    - `process_entity.path` object attribute
1. Added `extensions/windows/reg_key_path_t` as an Observable type -
`type_id: 45` and marked fields as this type
    - `reg_key.path` object attribute
    - `reg_value.path` object attribute

I have updated CHANGELOG.md with the same granularity of the information
it already uses.

##### Note On Compatibility

There is a failing compatibility validator, as I changed some fields
from `string_t` to the new `file_path_t` and `reg_key_path_t` - as this
seem to be the only way to mark several different fields as same
observable. Since both old and new types are strings (the new one just
adds semantics) I don't see how this could break type compatibility
(e.g. in generated classes). But I am happy to learn about any concerns
and also whether there is a better (more compatible) way to mark
multiple fields with the same observable type.

---------

Signed-off-by: Michal Dobisek <mdobisek@cisco.com>