Added Deleted to finding status_id enum (#1437)
#### Description of changes:
We have a few cases when we need to delete findings. For example:

1. The finding was created by misconfigured or buggy producer. When the
issues is discovered, the producer needs to inform all services
consuming the data that the finding is not valid anymore and should be
discarded.
2. A customer was evaluating the product, created some test findings and
now wants to start using the product in production. We need to delete
the data from POC.

Signed-off-by: Lukáš Křečan <lukas.krecan@sentinelone.com>