Added `status` to `related_event` object (#1434)
#### Description of changes:
We need to transfer status of related finding in `related_events` (the
same way severity is already transferred). There are two problems with
status.

1. The first one is technical. Finding status_id is declared directly in
finding class and can't be moved to dictionary as `status_id` is defined
in multiple places.
2. The other one is logical. Related event does not have to represent
only Findings, but also other type, so it's not clear what the enum
should contain.

In theory, we can introduce `status_id: int` but I don't see any value
in that (please prove me wrong)

Signed-off-by: Lukáš Křečan <lukas.krecan@sentinelone.com>
Signed-off-by: Lukáš Křečan <146833112+lukas-krecan-s1@users.noreply.github.com>