Handle Executable and Read-Only files in the File object (#1438)
#### Related Issue: #1418

#### Description of changes:
Added an `Executable File` enum value to the File object type_id enums.
Added an `is_read-only` attribute to the dictionary and redefined it
specifically in the `File` object.

After discussion, we felt that treating executable files as distinct
from regular files was warranted, vs. relying on the attributes of a
file by the file system. The rationale I used was that the actual
construction of the executable file is special, similar to a Symbolic
Link being special, or a Folder being special - as they are
distinguished currently in the enum list.

---------

Signed-off-by: Paul Agbabian <pagbabian@splunk.com>