Add `org` to `reporter` object (#1507)
#### Related PR:
#1476


#### Description of changes:

This PR simply adds the `org` object to the `reporter`, so to capture
the organization attributes pertaining to the entity from which the
event or finding was first reported.

NOTE: No `CHANGELOG` entry is needed for this, as the `reporter` object
was created within this same release, `1.7.0-dev`.

<img width="998" height="369" alt="image"
src="https://github.com/user-attachments/assets/75de168a-c259-4234-b602-a1036a3e545b"
/>

---------

Co-authored-by: Rajas <89877409+floydtree@users.noreply.github.com>
Co-authored-by: Ania Kacewicz <ania.kacewicz@gmail.com>