Added source and type to the metadata object. Updated the log_* attributes and descriptions for consistency. (#1483)
#### Related Issue: #1487

#### Description of changes:
Currently OCSF events don't explicitly call out the source or type of
the event, whether natively produced or mapped from another source. A
classic approach to high level classificaiton of telemetry data, or
experimental measurement data is to have standard `source` and `type`
labels. Although `metadata` has a general `labels` array, it is more
structured to standardize the names for these two dimensions of events.
`logger` and `loggers` are intended to indicate the chain of custody of
events in a pipeline. There is a special case where the "source" of an
event comes from a log.

Splunk has `host` `source` and `sourcetype` fields that correspond to
the physical origin of the event, (e.g. /var/log/messages) and its
combined source and format (e.g. cisco_syslog, Apache access_combined).
Given it is implicit that the format of an OCSF event is OCSF, `type`
within `metadata` is not a format but rather a classification (e.g.
firewall traffic logs).

These are optional attributes.

In the process of analyzing the gaps, I noticed that the `log_*`
attributes have evolved in an ambiguous way. Specifically it isn't clear
whether the log in question is a source log or a consumer log (e.g. a
SIEM). The `log_version` was an indirect format indicator but with no
detail. The `logger` object and `loggers` array were not clearly
distinguished from what is represented for `log_*` in `metadata`.

Hence I modified the descriptions in light of new, explicit `log_source`
and `log_format` attributes, and updated the description of
`log_version` which then entailed the deprecation of `version` from
`loggers` in favor of `log_version` to be consistent.

---------

Signed-off-by: Paul Agbabian <pagbabian@splunk.com>
Signed-off-by: Rajas <89877409+floydtree@users.noreply.github.com>
Co-authored-by: Rajas <89877409+floydtree@users.noreply.github.com>