Cross reference between `process` and  `win_service` objects. (#1526)
#### Related Issue: 

https://github.com/ocsf/ocsf-schema/issues/1525

#### Description of changes:

Add field to `win_service` object to refer to the process hosting it.
Add fields to `win_service` object so that the executable file, and the
optional dll file, can be fully described

Add array field to `process` to refer the the service(s) that the
process is currently hosting.

---------

Signed-off-by: Mitchell Wasson <4595268+mlmitch@users.noreply.github.com>
Co-authored-by: Jonathan Edwards <jonathae@cisco.com>
Co-authored-by: Mitchell Wasson <4595268+mlmitch@users.noreply.github.com>